+++ src/main/java/org/cnc/cncbot/dto/publicplayerinfo/Ew.java
@@ -18,6 +18,6 @@ private String n;
 
 private Integer r;
 
-private Integer ws;
+private Long ws;
 
 }
